Fix variable naming and fine-tune for ten epochs
Before, variable naming was inconsistent, leading to errors when
trying to run the code; these inconsistencies have been fixed. Also,
previously, fine-tuning the CNN was done for 5 epochs, which didn't
match the number of epochs the RNN was trained for, so we now
fine-tune for ten epochs to fix this inconsistency.